requirements.txt
# Core Framework
Django==5.0.13
djangorestframework==3.14.0

# Database (for production - PostgreSQL)
psycopg2-binary==2.9.9

# Web Server (for production)
gunicorn==21.2.0

# Static Files (for production)
whitenoise==6.6.0

# Utilities
python-decouple==3.8